+++ src/Faker.Tests/EducatorTests.cs
@@ -1,13 +1,20 @@
 ﻿using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RimuTec.Faker.Tests {
    [TestFixture]
    public class EducatorTests : FixtureBase {
+      [Test]
+      public void Campus_Happy_Days() {
+         // arrange
+
+         // act
+         var campus = Educator.Campus();
+
+         // assert
+         Assert.IsFalse(string.IsNullOrWhiteSpace(campus));
+         Assert.Greater(RegexMatchesCount(campus, @" "), 0);
+      }
+
       [Test]
       public void Course_Happy_Days() {
          // arrange