+++ src/Faker/Company.cs
@@ -12,7 +12,6 @@ namespace RimuTec.Faker {
    /// </summary>
    public static class Company {
       // Resources used by this class from https://github.com/stympy/faker/blob/master/lib/locales/en/company.yml
-      // [Manfred, 21jun2018]
 
       static Company() {
          const string yamlFileName = "RimuTec.Faker.locales.en.company.yml";