+++ releasenotes.md
@@ -1,7 +1,10 @@
 # Release Notes
 
 ## Version 0.24.0
+### Summary
+Educator is now on par with Ruby Faker for default locale ("en").
 ### New Features
+- Educator.Campus()
 - Educator.Course()
 - Educator.SecondarySchool()
 - Educator.University()